requirements.txt
cython
flup
flask
pyyaml
mwtemplates
oursql

